Simplify and add tests for token expiry methods
